IR: Fix the return value of DICompileUnit::getDWOId.
llvm-svn: 248341

+++ llvm/include/llvm/IR/DebugInfoMetadata.h
@@ -1026,7 +1026,7 @@ public:
   DIImportedEntityArray getImportedEntities() const {
     return cast_or_null<MDTuple>(getRawImportedEntities());
   }
-  unsigned getDWOId() const { return DWOId; }
+  uint64_t getDWOId() const { return DWOId; }
 
   MDString *getRawProducer() const { return getOperandAs<MDString>(1); }
   MDString *getRawFlags() const { return getOperandAs<MDString>(2); }
